💻 Simplify engine again - add Ordnance: a score that reflect the amount of powerups a ships holds - add a ordnance cap and associated error, prevent buy crate when ordnance is full - update matchmaking to use Ordnance - update currency rewards (3 for winner, 1 for looser, 1 for each when draw) (remove XP rewards) - remove stats (subsystems) - remove exotic modules

+++ programs/hologram/src/state/sub_types/wallet.rs
@@ -40,18 +40,18 @@ impl Wallet {
         Ok(())
     }
 
-    pub fn credit(&mut self, amount: u16, currency: Currency) -> Result<()> {
+    pub fn credit(&mut self, amount: u8, currency: Currency) -> Result<()> {
         match currency {
             Currency::ImperialCredit => {
                 self.imperial_credits = self
                     .imperial_credits
-                    .checked_add(amount)
+                    .checked_add(amount as u16)
                     .ok_or(HologramError::Overflow)?;
             }
             Currency::ActivateNanitePaste => {
                 self.activate_nanite_paste = self
                     .activate_nanite_paste
-                    .checked_add(amount)
+                    .checked_add(amount as u16)
                     .ok_or(HologramError::Overflow)?;
             }
         };
